refactor: world module
