+++ apps/server/src/context.ts
@@ -2,7 +2,7 @@ import type { Branded } from "@mp/std";
 import type { AuthToken, AuthServer } from "@mp/auth-server";
 import type { Logger } from "@mp/logger";
 import type { StateAccess } from "@mp/sync/server";
-import type { WorldState } from "./WorldState";
+import type { WorldState } from "./modules/world/WorldState";
 import type { ClientRegistry } from "./ClientRegistry";
 
 export interface ServerContext {
