+++ apps/server/src/package.ts
@@ -1,7 +1,7 @@
 // This is the file that represents what the @mp/server package exports.
 // This module should never export runtime code. Only use type exports.
 export type * from "./modules/character/schema";
-export type * from "./WorldState";
+export type * from "./modules/world/WorldState";
 export type * from "./context";
 export type * from "./modules/router";
 export type { ServerOptions } from "./serverOptions";
